+++ src/main/java/org/sopt/carena/healthreport/adapter/in/web/controller/HealthReportController.java
@@ -4,8 +4,8 @@ import org.sopt.carena.global.api.response.ApiResponse;
 import org.sopt.carena.global.api.response.SuccessResponse;
 import org.sopt.carena.healthreport.adapter.in.web.request.CreateHealthReportRequest;
 import org.sopt.carena.healthreport.adapter.in.web.code.SuccessCode;
-import org.sopt.carena.healthreport.application.dto.commend.CreateHealthReportCommand;
-import org.sopt.carena.healthreport.application.dto.commend.ExtractTextCommand;
+import org.sopt.carena.healthreport.application.dto.command.CreateHealthReportCommand;
+import org.sopt.carena.healthreport.application.dto.command.ExtractTextCommand;
 import org.sopt.carena.healthreport.application.dto.view.EntireHealthReportView;
 import org.sopt.carena.healthreport.application.dto.view.ExtractedTextView;
 import org.sopt.carena.healthreport.application.dto.view.HealthReportDateListView;
