+++ src/main/java/org/sopt/carena/healthreport/application/port/in/CreateHealthReportUseCase.java
@@ -1,6 +1,6 @@
 package org.sopt.carena.healthreport.application.port.in;
 
-import org.sopt.carena.healthreport.application.dto.commend.CreateHealthReportCommand;
+import org.sopt.carena.healthreport.application.dto.command.CreateHealthReportCommand;
 
 public interface CreateHealthReportUseCase {
 	void createHealthReport(CreateHealthReportCommand commend);
