+++ src/main/java/org/sopt/carena/healthreport/application/port/in/ExtractTextFromImageUseCase.java
@@ -1,6 +1,6 @@
 package org.sopt.carena.healthreport.application.port.in;
 
-import org.sopt.carena.healthreport.application.dto.commend.ExtractTextCommand;
+import org.sopt.carena.healthreport.application.dto.command.ExtractTextCommand;
 import org.sopt.carena.healthreport.application.dto.view.ExtractedTextView;
 
 public interface ExtractTextFromImageUseCase {
